Project1_Simple_REST_API/RestDjango/django_restframework/requirements.txt
asgiref==3.2.7
Django==3.0.5
djangorestframework==3.11.0
#pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.1
